Moved CSS class into `host` property
Since this is static class there is no reason to use a binding for it.
Firstly this change will result in very slight performace gain, since
there are less bindings to check on every change detection run. Secondly
it removes undesired public property, which will be exposed when
component is made compatible with `fullTemplateTypeCheck`
(see next commit).