Eliminate Angular component inheritance
It provides rather limited benefit in this case, but instead creates a
lot of time-consuming and hard to debug issues. For this reason the code
was inlined.

The specific issue is that Ivy wants to have selector-less @Directive()
on the base class, while VE is not happy with it. The published code
should be compiled for VE, so it can be later processed by ngcc, but
this fails on presence of @Directive(). If class does not have this
annotation, then it will be successfully compiled by VE, but will fail
on usage because without @Directive() annotations on the base class are
not recognized.

Fixes #154