Make MockFaIconLibrary throw on attempts to add icons or packs
Normally people should not add icons or packs outside of the
`AppModule`/`AppComponent` constructor, so this should only throw in
case of an error and help to detect a mistake (e.g. user imported
`FontAwesomeTestingModule`, but attempts to configure the regular
`FaIconLibrary` and is surprised that icons are not rendered properly).

Even if the above assumption turns out to be wrong it will be easier to
remove this error, then add it after the release (as it will be a
breaking change).