Fix ng add crash on Typescript 3.7.5
`angular-fontawesome` abuses `@schematics/angular` instead of re-implementing some AST manipulations. The problem is that source file is parsed by the workspace's TypeScript version (3.7.5 in this case), but `@schematics/angular` bundles TypeScript 3.8.3 and assumes that AST was produced by 3.8.3. The AST is incompatible, hence the crash.

The solution is to use TypeScript bundled with `@schematics/angular` to parse the source file instead of using workspace's version.

Fixes #237
Fixed #234